Dash/Blanket_Template/requirements.txt
# requirements.txt generated by rsconnect-python on 2021-05-03 16:58:33.628764
dash==1.19.0
dash-bootstrap-components==0.11.3
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
JayDeBeApi==1.2.3
numpy==1.20.1
pandas==1.2.3
paramiko==2.7.2
plotly==4.14.3
pymongo==3.11.3
python-dateutil==2.8.1
scikit-learn==0.24.1
scipy==1.6.1
statsmodels==0.12.2
Werkzeug==1.0.1
